Scratch Game
This feature was designed as a sales component for the Hubbard
Broadcasting websites.  There would be
three images: a winner, a loser and a cover.  The script getting the
games will do two random pulls one that will
pull the winning number and the second that will pull to see if the
winner will be pulled ( matching the second
randomly pulled number to the first randomly pulled number).  Two make
the odds easier, lower the second parameter in
the rand function.  Every three minutes, if the user has not played,
the game will reset itself due to a certain amount
winning in a given time.